### 2017-08-25 #379 @whedon
[PRE REVIEW]: Pyradigm: PYthon based data structure to improve Dataset's InteGrity in Machine learning workflows
**Submitting author:** @raamana (<a href="http://orcid.org/0000-0003-4662-0558">Pradeep Reddy Raamana</a>)
**Repository:** <a href="https://github.com/raamana/pyradigm" target ="_blank">https://github.com/raamana/pyradigm</a>
**Version:** v0.3.0.3
