### 2017-08-25 #379 @raamana
Thanks @arokem for editing this.

Two devs volunteered to review for JOSS that might fit the need here, would be @ahwillia and @stsievert. If @oesteban has time to review for another package, that'd be even better.
