+++ CONTRIBUTING.md
@@ -1,0 +1,18 @@
+Hello there! 
+
+Thank you for the checking out `pyradigm`, and for your interest in contributing. 
+
+You could contribute in the areas, including but not limited to:
+ * identifying any bugs by focusing on the areas of your expertise 
+    * write unit tests or property tests to assert its advertised functionality
+    * write more examples (esp. the ones for your usecase)
+ * installing and using it to ensure it works as advertised (and fix it otherwise)
+ * helping improving python version compatibility (2.6 to 3.6)
+ * sharpening or improving the focus of the repo in general
+ * improving, clarifying and adding documentation
+ * propose interesting, popular and exciting use cases that this package could support to further multiply its value
+ * improving code base in terms of modularity
+   * ideas for integration with other use cases
+  
+  
+ *Under construction*
