### 2017-08-25 ddcad06
Adding code quality badge [skip ci]
