### 2017-08-26 #379 @stsievert
@raamana I'd be happy to review this!
