### 2017-08-26 #379 @ahwillia
I could give a brief review as well. From a quick look over - I'd be interested in an explanation about how well this plays with pandas and other existing tools/libraries.
