### 2017-08-28 #379 @arokem
Thanks @raamana for the submission. Looks interesting!

Thanks @ahwillia and @stsievert for agreeing to review this!

I'd be happy if both of you could review this. I will try to assign both of you as reviewers through our bot, @whedon, but I am not sure whether that would work automatically. Either way, please consider both of you as reviewers for this, and I will wait for both of your 👍 before accepting this paper.
