### 2017-08-28 #379 @arokem
Alright, @stsievert, @ahwillia : please head over to #382 for the review.

I will close this one.
