### 2017-08-28 #379 @stsievert
Looks like you have a typo in https://github.com/openjournals/joss-reviews/issues/379#issuecomment-325241673 (an extra "e" on assign)
