### 2017-08-28 #382 @ahwillia
@arokem - is this the appropriate place for me to ask any questions to you?

At the moment, I'm wondering if I should limit the review narrowly to the checkboxed above or if more long form suggestions are welcome. For example, I think it would be nice if this package was compatible with Python 3+, and it shouldn't take that much effort to make that happen. Is that a reasonable request from the standpoint of JOSS? I feel a bit odd opening issues for these sorts of suggestions on the target repo.

Thanks for walking me through this.
