### 2017-08-28 #382 @raamana
Hi Alex,

Thank you for your time to review this and the positive support for the ideas and motivation for this project. Much appreciate it. I'm looking forward see your ideas and suggestions for how to improve its value and interoperability with other tools and packages.

Support for python 3+ is certainly on my TODO list, perhaps I should note it on the repo itself in its roadmap. It's just limited by my time or contributions from others interested.

In fact, I already have given it a try few weeks ago, either it's not so simple or I need to up my python module packaging and version compatibility game. If you have clear suggestions, I'll happily follow the suggestions, provided they wouldn't need take way too long.

It's a great suggestion for improvement, but I'm not sure if that's a requirement for acceptance. Perhaps @arokem can clarify.
