### 2017-08-28 #382 @ahwillia
Yep - I'm not a fan of requirements and reviewer gate keeping, but was trying to figure out the scope of things acceptable to comment on and suggest. I guess my question is whether a repo is "accepted" once I click all the boxes, or if I am expected to also make high-level suggestions and criticisms that the editor decides are either necessary or not.

Shame to hear that Python3 wasn't straightforward to implement. But that's a minor comment anyways.

I'll try to block off an afternoon later this week to do the review in full. I'm just trying to feel out exactly what has to be done so I can finish it in one go.
