### 2017-08-28 #382 @raamana
To my understanding, @arokem can correct me if I am wrong, if you tick all the questions off (implying the questions therein are answered), I believe it will be accepted. However, the key section for your concerns in Functionality - where "Have the functional claims of the software been confirmed?" touches on the Python 3 compatibility. If pyradigm were to claim compatibility for python 3+ (which it doesn't) and fail to satisfy that, then it needs to be addressed.
