### 2017-08-28 #382 @raamana
I just updated the README to clarify this.
