### 2017-08-28 raamana/pyradigm#2 @stsievert
Improve documentation
It would be useful to describe each function more completely, by documenting the inputs args and the return values. This will let people use your software more easily, and it would be the first thing I would look for.

A useful guideline to follow is probably the [numpy documentation standards][1]. Tools like [Sphinx] (with [numpydoc]) can autogenerate the documentation and put it online.

For example, the documentation for [MLDataset.add_sample][2] is currently

``` python
def add_sample(self, sample_id, features, label, class_id=None, feature_names=None):
    """Adds a new sample to the dataset with its features, label and class ID.
    This is the preferred way to construct the dataset."""
```

and following the NumPy doc standards would be something like

``` python
def add_sample(self, sample_id, features, label, class_id=None, feature_names=None):
    """Adds a new sample to the dataset with its features, label and class ID.

    This is the preferred way to construct the dataset.

    Parameters
    ----------
    sample_id : str, int
        The identifier that uniquely identifies this sample.
    features : list, ndarray
        The features for this example
    label : int, str
        The label for this example.
    class_id : int, str
        The class for this example
    feature_names : list
        The names for each feature. Assumed to be in the same order as `features`

    Raises
    ------
    ValueError
        If `sample_id` is already in the MLDataset

    """
```

(numpydoc used because that's what I'm familiar with)

[Sphinx]:http://www.sphinx-doc.org
[numpydoc]:https://github.com/numpy/numpydoc
[1]:https://github.com/numpy/numpy/blob/master/doc/HOWTO_DOCUMENT.rst.txt
[2]:https://github.com/raamana/pyradigm/blob/9a168079b0ef6409d478bbeff8d47bde046a08d5/pyradigm/pyradigm.py#L184
