### 2017-08-28 #382 @stsievert
Thanks for submitting this @raamana! Most of this looks good, and +1 for positive and negative tests.

I do have a couple comments after a quick review this morning:

* Documentation: I'd like to see better documentation, and have filed https://github.com/raamana/pyradigm/issues/2
* Statement of need: could you include a paragraph about how this extends pandas? Would including pandas simplify the implementation?
* Community guidelines: CONTRIBUTING.md describes what to contribute but does not describe *how* and misses the points mentioned in https://github.com/openjournals/joss-reviews/issues/382#issue-253207360. Could you clarify this?

I'll try to put in more time for review later this week.
