### 2017-08-28 raamana/pyradigm#2 @raamana
Hey Scott, thanks for the detailed comment.  This is definitely at the top of my plan. I will wait for comments from all the reviewers, esp. regarding the API and inter-operability etc, to finalize the API, and document the more or less stable API thoroughly.
