### 2017-08-28 f92d6dd
adding guidelines for contributing [skip ci]
