+++ CONTRIBUTING.md
@@ -1,9 +1,13 @@
 Hello there! 
 
-Thank you for the checking out `pyradigm`, and for your interest in contributing. 
+Thank you for the checking out `pyradigm`, and for your interest in contributing. I appreciate it.
+
+Please check the following guidelines for contributing:
+http://contribution-guide-org.readthedocs.io/
 
 You could contribute in the areas, including but not limited to:
- * identifying any bugs by focusing on the areas of your expertise 
+ * sharing this package to those interested
+ * using it which helps in identifying any bugs by focusing on the areas of your expertise 
     * write unit tests or property tests to assert its advertised functionality
     * write more examples (esp. the ones for your usecase)
  * installing and using it to ensure it works as advertised (and fix it otherwise)
@@ -13,6 +17,8 @@ You could contribute in the areas, including but not limited to:
  * propose interesting, popular and exciting use cases that this package could support to further multiply its value
  * improving code base in terms of modularity
    * ideas for integration with other use cases
+ * Anything else I may have missed and you think is necessary, important and just plain interesting! :)
   
+Thank you again for you interest. Appreciate your contribution in all its forms.
   
  *Under construction*
