### 2017-08-28 #382 @raamana
Thanks Scott. Full docs is definitely at the top of my plan. I will wait for comments from all the reviewers, esp. regarding the API and interoperability etc, to finalize the API, and document it thoroughly.

Let me answer the next two points shortly.
