### 2017-08-28 72a6450
comment on relation to pandas [skip ci]
