### 2017-08-28 #382 @raamana
So I have updated the README  and contribution guidelines as @stsievert  suggested, take a look. Perhaps you can tick off the issues for the review as you see them done/to-be-done.

I will wait for @ahwillia's comments (and Scott's if any) before starting to document the API - this is one area all of your comments will be helpful.
