### 2017-08-28 1e4c31c
elaborating on motivation [skip ci]
