### 2017-08-28 raamana/pyradigm#2 @stsievert
Keep in mind that you can get free web hosting through GitHub by publishing the the gh-pages branch (which is then available at [username].github.io/[repo]
