### 2017-08-28 3b9ae56
Clean __contains__
