### 2017-09-04 #382 @arokem
Hi everyone!

@ahwillia : this is the place to ask me questions, though as you can see, it might take me a little while to answer...

At any rate, thanks for starting the discussion here. Let me see if I can answer the questions that have come up so far: extending support to Python 3 is not a requirement for publication, but I do agree that it would improve the software, so worth consider doing now that you are responding to a lot of external reviews.

On the other hand, documentation is a requirement, so you would have to address https://github.com/raamana/pyradigm/issues/2.

Thanks again, and please keep me in the loop as you go through the requirements. Either @ahwillia and @stsievert can tick the boxes above, but I will ask that you both also explicitly approve the submission at the end of the process.
