### 2017-09-04 #382 @ahwillia
@raamana - I'm okay to sign off on this once you address @stsievert's comments. Fixing up the documentation and giving a clear statement of purpose (with comparison to pandas and cross-validation tools in scikit-learn) are the top priority issues.

You can treat everything below as a suggestion.

* [The example](https://github.com/raamana/pyradigm/blob/master/PyradigmExample.ipynb) notebook is good but a bit long. It would be really helpful to new users if you could compartmentalize the functionality. In particular, the table of contents could have more direct headings, such as:
    * *Constructing a dataset*
    * *Cross-validation*
    * *Saving/reloading a dataset*
    * *Exporting to numpy*
    * etc.

* Is "dataset" the right name? What about dataframe or datatable? This is what pandas / pytables call them...

* Saving datasets in HDF5 or [Feather](https://github.com/wesm/feather) format would be really helpful. I am always looking for a better way to save and reload datasets, because I'm too impatient to learn [h5py](http://www.h5py.org/). Pickling has [well-documented issues](http://www.benfrederickson.com/dont-pickle-your-data/), so packages that make these other formats easily accessible are a big plus.

* It seems like the user needs export their data using the `dataset.data_and_labels()` command in order to use scikit-learn? It seems like that slightly defeats the purpose of the package. Since scikit-learn has a very strict API, would it make sense to fit models internally within the dataset object? Something like:

```python
from sklearn import svm
dataset.fit_model(svm.SVC(gamma=0.001, C=100.))
Y = dataset.predict()
 ```

I could see this getting tricky and beyond the scope of the current submission (e.g. what happens if you fit multiple models? Do you just delete the last one). Anyways, just a thought I had.
