### 2017-09-04 #382 @raamana
Thanks Alex - very useful comments.

Except for detailed API documentation, I think I've already addressed @stsievert's comments. I think he wants to take another pass to review this.

 * Your suggestions are very good, esp. making example TOC headings more direct and simple to read :). I changed the headings for now, will revamp once I get more time, with more later examples (showing how it is used in [neuropredict](https://github.com/raamana/neuropredict) for example).
 * dataset is the right name in my own mind, but I agree with you it can other meanings in other domains. Calling it a DataFrame (as in pandas) collapse its meaning to 2D object (which it currently is), but I'd like to try convey it can store arbitrary data with trivial changes
* Good suggestions for serializations - will consider the suggestions, but from research 6 months ago, pickle serves the purpose (for all its flaws) for now, if users don't do stupid things. json is certainly is a good alternative , but its disk size is higher (don't know how much) than pickle's binary format. This can be a problem for some people (like me), who uses a large number of datasets for each project. Improving this is certainly on the roadmap.
* Yes allowing model fitting will open another can of worms :).. [neuropredict](https://github.com/raamana/neuropredict) serves that purpose.
