### 2017-09-04 #382 @raamana
One thing I'd like @stsievert and @ahwillia to really give it a serious thought is on API. How can we improve it to ease your ML workflow on a daily/weekly basis?

I will get back to you soon with improved API docs
