### 2017-09-04 #382 @ahwillia
> How can we improve API to ease your ML workflow on a daily/weekly basis?

I don't think I'm the right person to ask, as I'm not used to dealing with datasets with very cumbersome metadata. Honestly, I don't even use pandas because working directly with numpy arrays serves my purposes. I also end up using a lot of nested dicts, which is maybe suboptimal, but this package doesn't appear to address that issue.
