### 2017-09-04 1a1bbc4
Pass additional kwargs to pickle.load (in __load)
