### 2017-09-04 #382 @stsievert
> How can we improve API to ease your ML workflow on a daily/weekly basis?

Possible API improvements:

* you have implemented `__contains__`. If `i in dataset`, I would expect `dataset[i]` return the value.
    * could be implemented with `__get_subset_from_dict`
    * could simplify `get_subset`
* `glance`: this is similar to Pandas `df.head`. Rename?
* `summarize_classes` is similar to Pandas `df.describe`. Rename?
* Implementation of `__del__` could be used to delete a sample
    * could be implemented with `del_sample`
* In your implementation of `num_features`, you raise a warning if you set. Make it a function instead? You can not set a function value.
* You have some functions that could be marshaled to a utils file. The functions below use none of the class features, and I don't know if they're appropriate for the class `MLDataset`.
    * `__take`
    * `__str_names`
    * `keys_with_value`

These are all first-glance comments. Let me know how any of these comments can be improved.
