### 2017-09-04 raamana/pyradigm#6 @stsievert
Python 3 ImportError
I installed with `pip install pyradigm`. Then with Python 3.6,

``` python
In [2]: import pyradigm
ImportError                               Traceback (most recent call last)
<ipython-input-2-1e91a6f90dae> in <module>()
----> 1 import pyradigm

/Users/scott/anaconda/lib/python3.6/site-packages/pyradigm/__init__.py in <module>()
      1
----> 2 from pyradigm import MLDataset
      3
      4 __all__ = [ 'pyradigm', 'MLDataset' ]
      5

ImportError: cannot import name 'MLDataset'
```

But under Python 2.7:

``` python
In [1]: from pyradigm import MLDataset

In [2]:
```
