### 2017-09-04 #382 @stsievert
I've gone through the checklist, and here are my findings (I can't edit the top comment)

- [x] repository
- [x] license (MIT)
- [ ] version. The version on PyPi matches but there are no GitHub releases.
- [x] authorship

- [ ] installation. Issue https://github.com/raamana/pyradigm/issues/6 filed.
- [ ] functionality. I need to test the last cell in [PyradigmExample.ipynb] about sklearn support.
- [x] statement of need. There is one on the readme
    * could this library somehow wrap Pandas?
- [x] installation instructions with PyPI.
- [x] Example usage with [PyradigmExample.ipynb]
- [ ] functionality documentation. Needs docs.
- [x] automated tests.
- [x] Community guidelines
- [x] authors
- [x] statement of need. There is one present.
- [ ] references. I don't believe you should self-reference your paper; other papers don't. Maybe @arokem can chime in?

[PyradigmExample.ipynb]:https://github.com/raamana/pyradigm/blob/master/PyradigmExample.ipynb
