### 2017-09-04 raamana/pyradigm#6 @raamana
Thanks Scott, I will improve the compatibility with py3+ soon, but should this hold back the acceptance? I think Ariel clarified it was encouraged, but not necessary.
