### 2017-09-04 22a9c38
python 3 compatibility
