+++ .travis.yml
@@ -1,7 +1,12 @@
 language: python
+cache: pip
 python:
   - "2.7"
+  - "3.5"
+  - "3.6"
+
 # command to install dependencies
 install: "pip install -r requirements.txt"
+
 # command to run tests
 script: pytest
