### 2017-09-05 2c35f11
more python 3 compatibility issues resolved
