### 2017-09-05 raamana/pyradigm#6 @raamana
Python 3.5, 3.6 and 2.7 are all supported now - with 2c35f1154976af7dd40f1a4914ebc330e4f75cb2.

check https://travis-ci.org/raamana/pyradigm

Reopen if something is amiss.
