### 2017-09-05 #382 @raamana
Glad to update you guys that paradigm works on both 2.7 and 3.6, along with 3.5: https://travis-ci.org/raamana/pyradigm
