### 2017-09-05 bbc638d
updating version numbers and fixing authors [skip ci]
