+++ setup.py
@@ -4,7 +4,7 @@ import os
 from setuptools import setup, find_packages
 
 setup(name='pyradigm',
-      version='0.3.0.3',
+      version='0.4',
       description='Python-based data structure to improve handling of datasets in machine learning workflows',
       long_description='Pyradigm: Python-based data structure to improve handling of datasets in machine learning workflows',
       keywords='machine learning, test dataset, python, workflow, provenance, data structure',
@@ -13,5 +13,16 @@ setup(name='pyradigm',
       url='https://github.com/raamana/pyradigm',
       packages=find_packages(exclude=["*.tests", "*.tests.*", "tests.*", "tests"]), # packages=['pyradigm'],
       install_requires=['numpy', 'setuptools'],
+      classifiers=[
+            'Intended Audience :: Science/Research',
+            'Programming Language :: Python',
+            'Topic :: Software Development',
+            'Topic :: Scientific/Engineering',
+            'Operating System :: Microsoft :: Windows',
+            'Operating System :: POSIX',
+            'Operating System :: Unix',
+            'Operating System :: MacOS',
+            'Programming Language :: Python :: 2.7',
+            'Programming Language :: Python :: 3.6',
+      ],
      )
-
