### 2017-09-05 a55ef47
documenting every single input and error/warning to be raised.
