+++ pyradigm/pyradigm.py
@@ -72,6 +72,15 @@ class MLDataset(object):
 
         Also returns sample_ids in the same order.
 
+        Returns
+        -------
+        data_matrix : ndarray
+            2D array of shape [num_samples, num_features] with features  corresponding row-wise to sample_ids
+        labels : ndarray
+            Array of numeric labels for each sample corresponding row-wise to sample_ids
+        sample_ids : list
+            List of sample ids
+
         """
 
         sample_ids = np.array(self.keys)
@@ -167,12 +176,38 @@ class MLDataset(object):
         return np.array(['f{}'.format(x) for x in range(num)])
 
     def glance(self, nitems=5):
-        """Quick and partial glance of the data matrix."""
+        """Quick and partial glance of the data matrix.
+
+        Parameters
+        ----------
+        nitems : int
+            Number of items to glance from the dataset.
+            Default : 5
+
+        Returns
+        -------
+        dict
+
+        """
         nitems = max([1, min([nitems, self.num_samples-1])])
         return self.__take(nitems, iter(self.__data.items()))
 
     def summarize_classes(self):
-        "Summary of classes: names, numeric labels and sizes"
+        """
+        Summary of classes: names, numeric labels and sizes
+
+        Returns
+        -------
+        tuple : class_set, label_set, class_sizes
+
+        class_set : list
+            List of names of all the classes
+        label_set : list
+            Label for each class in class_set
+        class_sizes : list
+            Size of each class (number of samples)
+
+        """
 
         class_sizes = np.zeros(len(self.class_set))
         for idx, cls in enumerate(self.class_set):
@@ -182,8 +217,35 @@ class MLDataset(object):
 
     # TODO try implementing based on pandas
     def add_sample(self, sample_id, features, label, class_id=None, feature_names=None):
-        """Adds a new sample to the dataset with its features, label and class ID.
-        This is the preferred way to construct the dataset."""
+        """
+        Adds a new sample to the dataset with its features, label and class ID.
+
+        This is the preferred way to construct the dataset.
+
+        Parameters
+        ----------
+        sample_id : str, int
+            The identifier that uniquely identifies this sample.
+        features : list, ndarray
+            The features for this sample
+        label : int, str
+            The label for this sample
+        class_id : int, str
+            The class for this sample
+        feature_names : list
+            The names for each feature. Assumed to be in the same order as `features`
+
+        Raises
+        ------
+        ValueError
+            If `sample_id` is already in the MLDataset, or
+            If dimensionality of the current sample does not match the current, or
+            If `feature_names` do not match existing names
+        TypeError
+            If sample to be added is of different data type compared to existing samples.
+
+        """
+
         if sample_id not in self.__data:
             if self.num_samples <= 0:
                 self.__data[sample_id] = features
@@ -210,13 +272,26 @@ class MLDataset(object):
                         self.__feature_names = np.array(feature_names)
                     else: # if set already, ensure a match
                         print('')
-                        assert np.array_equal(self.feature_names, np.array(feature_names)), \
-                            "supplied feature names do not match the existing names!"
+                        if not np.array_equal(self.feature_names, np.array(feature_names)):
+                            raise ValueError("supplied feature names do not match the existing names!")
         else:
             raise ValueError('{} already exists in this dataset!'.format(sample_id))
 
     def del_sample(self, sample_id):
-        """Method remove a sample from the dataset."""
+        """
+        Method remove a sample from the dataset.
+
+        Parameters
+        ----------
+        sample_id : str
+            sample id to be removed.
+
+        Raises
+        ------
+        UserWarning
[…]
