### 2017-09-05 raamana/pyradigm#6 @stsievert
Python 3 support is not required, but if it doesn't install under Python 3 there should be a note in the readme.

In __init__.py I don't think you need to specify exactly 2.7.13. 2.7 is good enough.
