### 2017-09-05 06d0113
adding docstring for the copy constructor and others missed [skip ci]
