### 2017-09-05 raamana/pyradigm#6 @raamana
yes, 2.7 might be enough, but I wanted to be specific as to what was being tested for support.
