### 2017-09-05 raamana/pyradigm#2 @raamana
Thanks for the tip of gh-pages. I know of it [I think I did create one before] and readthedocs also, but they are not trivial to get done or to maintain. Once more and more people use this or contribute here, I can invest more time.

btw, 06d0113a454e13dfb869b9a948010da8537f444e should address the docs requirement.
