### 2017-09-05 #382 @raamana
- [x] Full docs for API
- [x] same version everywhere
- [x] installations tests on CI
- [x] functionality tests on CI

Thanks again. Let me know if you guys need anything else to close the review.
