### 2017-09-05 raamana/pyradigm#6 @stsievert
Can you be specific in other places, and not enforce that the version must be 2.7.13?

> the revision number is incremented when minor bugs are fixed. https://en.m.wikipedia.org/wiki/Software_versioning

So 2.7.13 should only have minor bug fixes, which is true according to the [2.7.13 changelog][1].

[1]:https://hg.python.org/cpython/raw-file/v2.7.13/Misc/NEWS
