### 2017-09-05 #382 @raamana
Scott, thanks for the suggestions to improve the API. I will keep them in mind and improve the API in due course. Note the API must cater to non-expert programmers/users from neuro- or related fields (who know of datasets/samples etc, but never heard of pandas). I'm not targeting those who already know of or use pandas :)
