### 2017-09-05 007a17a
turning num_features setter to static method [skip ci]
