### 2017-09-05 dcf5f2d
code clean up [skip ci]
