+++ pyradigm/pyradigm.py
@@ -1,11 +1,13 @@
-import numpy as np
-from collections import Counter, OrderedDict, Sequence
-from itertools import takewhile, islice
-import random
-import warnings
+import copy
 import os
 import pickle
-import copy
+import random
+import warnings
+from collections import Counter, OrderedDict, Sequence
+from itertools import islice
+
+import numpy as np
+
 
 # TODO profile the class for different scales of samples and features
 class MLDataset(object):
@@ -13,7 +15,7 @@ class MLDataset(object):
 
     def __init__(self, filepath=None, in_dataset=None,
                  data=None, labels=None, classes=None,
-                 description='', feature_names = None):
+                 description='', feature_names=None):
         """
         Default constructor.
         Suggested way to construct the dataset is via add_sample method, one sample at a time.
@@ -239,7 +241,7 @@ class MLDataset(object):
         dict
 
         """
-        nitems = max([1, min([nitems, self.num_samples-1])])
+        nitems = max([1, min([nitems, self.num_samples - 1])])
         return self.__take(nitems, iter(self.__data.items()))
 
     def summarize_classes(self):
@@ -320,7 +322,7 @@ class MLDataset(object):
                     # class gets here when adding the first sample, after dataset was initialized with empty constructor
                     if self.__feature_names is None:
                         self.__feature_names = np.array(feature_names)
-                    else: # if set already, ensure a match
+                    else:  # if set already, ensure a match
                         print('')
                         if not np.array_equal(self.feature_names, np.array(feature_names)):
                             raise ValueError("supplied feature names do not match the existing names!")
@@ -374,14 +376,14 @@ class MLDataset(object):
         subset_idx = np.asarray(subset_idx)
         if not (max(subset_idx) < self.__num_features) and (min(subset_idx) >= 0):
             raise UnboundLocalError('indices out of range for the dataset. '
-                              'Max index: {} Min index : 0'.format(self.__num_features))
+                                    'Max index: {} Min index : 0'.format(self.__num_features))
 
         sub_data = {sample: features[subset_idx] for sample, features in self.__data.items()}
         new_descr = 'Subset features derived from: \n ' + self.__description
         subdataset = MLDataset(data=sub_data,
                                labels=self.__labels, classes=self.__classes,
                                description=new_descr,
-                               feature_names= self.__feature_names[subset_idx])
+                               feature_names=self.__feature_names[subset_idx])
 
         return subdataset
 
@@ -420,7 +422,7 @@ class MLDataset(object):
             class_ids = class_id
 
         non_existent = set(self.class_set).intersection(set(class_ids))
-        if len(non_existent)<1:
+        if len(non_existent) < 1:
             raise ValueError('These classes {} do not exist in this dataset.'.format(non_existent))
 
         subsets = list()
@@ -431,7 +433,7 @@ class MLDataset(object):
 
         return self.get_subset(subsets)
 
-    def train_test_split_ids(self, train_perc = None, count_per_class = None):
+    def train_test_split_ids(self, train_perc=None, count_per_class=None):
         """
         Returns two disjoint sets of sample ids for use in cross-validation.
 
@@ -463,15 +465,15 @@ class MLDataset(object):
 
         """
 
-        _, _, class_sizes = self.summarize_classes()
+        _ignore1, _ignore2, class_sizes = self.summarize_classes()
         smallest_class_size = np.min(class_sizes)
 
-        if count_per_class is None and (train_perc>0.0 and train_perc<1.0):
+        if count_per_class is None and (0.0 < train_perc < 1.0):
             if train_perc < 1.0 / smallest_class_size:
                 raise ValueError('Training percentage selected too low '
                                  'to return even one sample from the smallest class!')
[…]
