### 2017-09-05 ad5f661
converting assert statements to if statements
