### 2017-09-05 9bbc7d7
Add Codacy badge
