+++ README.md
@@ -3,6 +3,7 @@
 [![status](http://joss.theoj.org/papers/c5c231486d699bca982ca7ebd9cf32d2/status.svg)](http://joss.theoj.org/papers/c5c231486d699bca982ca7ebd9cf32d2)
 [![travis](https://travis-ci.org/raamana/pyradigm.svg?branch=master)](https://travis-ci.org/raamana/pyradigm.svg?branch=master)
 [![Code Health](https://landscape.io/github/raamana/pyradigm/master/landscape.svg?style=flat)](https://landscape.io/github/raamana/pyradigm/master)
+[![Codacy Badge](https://api.codacy.com/project/badge/Grade/cffd80f290544e2e824011bfccf35ff8)](https://www.codacy.com/app/raamana/pyradigm?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=raamana/pyradigm&amp;utm_campaign=Badge_Grade)
 [![PyPI version](https://badge.fury.io/py/pyradigm.svg)](https://badge.fury.io/py/pyradigm)
 [![codecov](https://codecov.io/gh/raamana/pyradigm/branch/master/graph/badge.svg)](https://codecov.io/gh/raamana/pyradigm)
 
