### 2017-09-06 #382 @arokem
@stsievert : IIUC - the reference in the paper is to a different, more neuro-specific project. That's fine.

@raamana : Regarding documentation of the software -- I don't think that the current documentation is sufficient. I would recommend adding a sphinx documentation website, including full reference documentation of the package and its API. If you want to automate documentation building on your CI, this is a good option: https://drdoctr.github.io/doctr/
