### 2017-09-06 #382 @stsievert
Thanks for keeping the API comments in mind!

> the reference in the paper is [...] different

Whoops -- that's what you get for skimming the URL.
