+++ docs/examples.rst
@@ -1,0 +1,8 @@
+Usage
+-----
+
+A tutorial-like example is given in the following Jupyter notebook:
+
+`Pyradigm Example <https://github.com/raamana/pyradigm/blob/master/PyradigmExample.ipynb>`_.
+
+Please feel free to reach out to me if something is unclear or missing.
