+++ docs/usage.rst
@@ -1,0 +1,1365 @@
+--------
+Usage
+--------
+
+A tutorial-like example is given in the following Jupyter notebook:
+
+`Pyradigm Example <https://github.com/raamana/pyradigm/blob/master/PyradigmExample.ipynb>`_,
+
+which is reproduced here for your convenience.
+
+Table of Contents
+
+-  Motivation
+
+-  Constructing a dataset
+
+-  Attributes
+
+-  Iteration over samples
+
+-  Subset selection
+
+-  Saving/reloading a dataset (Serialization)
+
+-   Combining multiple datasets and arithmetic on useful subsets within
+   datasets
+
+-  Exporting to numpy and portability (e.g. with sklearn)
+
+
+
+Illustration of Pyradigm's utility via examples and their use-cases
+-------------------------------------------------------------------
+
+This class is greatly suited for neuroimaging applications (or any other
+domain), where each sample needs to be uniquely identified with a
+subject ID (or something similar).
+
+Key-level correspondence across data, labels (1 or 2), classnames
+('healthy', 'disease') and the related helps maintain data integrity and
+improve the provenance, in addition to enabling traceback to original
+sources from where the features have been originally derived.
+
+Just to given you a concrete examples, let's look at how an ML dataset
+is handled traditionally.
+
+You have a matrix X of size n x p, with n samples and p features, and a
+vector y containing the target values (or class labels or class
+identifiers). This X and y serves as training (and test set) for a
+classifier like SVM to fit the data X to match y as accurately as
+possible.
+
+Let's get a little more concrete:
+
+.. code-block:: python
+
+    import sys, os
+    import numpy as np
+    import matplotlib
+    %matplotlib
+    %matplotlib inline
+    import matplotlib.pyplot as plt
+    
+    n = 10 # number of samples
+    p = 3  # number of features
+    
+    X = np.random.random([n, p]) # random data for illustration
+    y = [1]*5 + [2]*5            # random labels ...
+    
+    np.set_printoptions(precision=2) # save some screen space
+    print 'X : \n', X
+    print 'y : \n', y
+
+
+.. code-block:: python
+
+    Using matplotlib backend: TkAgg
+    X : 
+    [[ 0.64  0.48  0.88]
+     [ 0.19  0.05  0.12]
+     [ 0.13  0.1   0.68]
+     [ 0.99  0.19  0.39]
+     [ 0.86  0.36  0.91]
+     [ 0.83  0.98  0.32]
+     [ 0.86  0.35  0.3 ]
+     [ 0.32  0.65  0.83]
+     [ 0.6   0.75  0.53]
+     [ 0.12  0.52  0.41]]
+    y : 
+    [1, 1, 1, 1, 1, 2, 2, 2, 2, 2]
+
+
+Almost all the machine learning toolboxes take their input in this form:
+X and y, regardless of the original source that produced these features
+in the first place.
+
+This is all fine if all you ever wanted to do is to extract some
+features, do some machine learning and dispose these features away!
+
+*\* But this is almost never the case!*\ \*
+
+Because it doesn't simply end there.
+
+At a minimum, I often need to know \* which samples are misclassified -
+meaning you need to know what the identifiers are and not simply their
+row indices in X? \* what are the charecteristics of those samples? \*
+what classes do they belong to?
+
+And all this info needs to be obtained \* without having to write lots
+of code connecting few non-obvious links to disparate sources of data
[…]
