+++ docs/API.rst
@@ -1,0 +1,10 @@
+
+API Reference
+--------------
+
+A tutorial-like presentation is available at :doc:`usage`, using the following API.
+
+.. automodule:: pyradigm
+   :members:
+   :undoc-members:
+   :show-inheritance:
