+++ docs/installation.rst
@@ -1,0 +1,18 @@
+------------
+Installation
+------------
+
+Pyradigm can easily be installed with a single command:
+
+``pip install pyradigm``
+
+
+If you lack ``sudo`` access, try
+
+``pip install pyradigm --user``
+
+Requirements
+------------
+
+ * Packages: ``numpy``
+ * Supported versions: 2.7, 3.5 and 3.6

[…]
