### 2017-09-06 a5e88b0
helping RTD builds with numpy C requirements [skip ci]
