### 2017-09-06 1f2b1a6
still fixing RTD builds [skip ci]
