+++ conf.py
@@ -30,7 +30,7 @@ class Mock(MagicMock):
     def __getattr__(cls, name):
             return MagicMock()
 
-MOCK_MODULES = [ 'argparse', 'numpy']
+MOCK_MODULES = [ 'argparse', 'numpy', 'numpydoc']
 sys.modules.update((mod_name, Mock()) for mod_name in MOCK_MODULES)
 
 # -- General configuration ------------------------------------------------
