### 2017-09-06 #382 @raamana
I've finally tamed Sphinx to help me produce the docs : http://pyradigm.readthedocs.io/en/latest/

Thanks @arokem for the doctr tip, and I will try to use that too.

Is there anything else I need to do for the acceptance?
