### 2017-09-07 #382 @raamana
An older theme showed a nice table of attributes and methods at the top, using autosummary directive, which doesnt seem to work with the current theme or docs hosting service readthedocs. If you can help get that to work, that'll be appreciated.
