### 2017-09-08 #382 @arokem
Where does https://github.com/raamana/pyradigm/pull/8 stand?

Is this required for acceptance, @stsievert? Does your 👍 on previous comment mean that it's ready to go from your point of view?

@raamana : for the table, you mean something like this?

 http://yeatmanlab.github.io/pyAFQ/reference/AFQ.html

Sorry - I am not exactly sure how that comes about. But feel free to copy the sphinx configuration from that project!
