### 2017-09-08 #382 @raamana
I just merged the PR https://github.com/raamana/pyradigm/pull/8

@arokem, yes. I gotten it for the older theme, but the shiny new (mobile-friendly) RTD theme doesn't seem to interact well with the autosummary directive. Its minor anyways - I might make a manual table later on, it would be ideal to have it automated. I will check their conf.
