### 2017-09-08 #382 @stsievert
> Is this required for acceptance, @stsievert? Does your 👍 on previous comment mean that it's ready to go from your point of view?

:+1: for review here -- I've checked all the boxes and am satisfied! In https://github.com/openjournals/joss-reviews/issues/382#issuecomment-327631871 I was giving a :+1: for documentation, not for review, but I am satisfied now.
