### 2017-09-08 #382 @raamana
Awesome. Thanks Alex, Scott and Ariel 👍, much appreciate your time and effort.

Let's hope pyradigm will find some users and the users will find it useful :)
