### 2017-09-09 6e36db7
setting the class label even when not supplied by user;
