+++ conf.py
@@ -94,6 +94,8 @@ pygments_style = 'sphinx'
 todo_include_todos = False
 
 
+autosummary_generate = True
+
 # -- Options for HTML output ----------------------------------------------
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
