### 2017-09-11 #382 @arokem
@arfon: I believe this paper is publishable in its current form. Please let me know if we need to do anything else here.
