### 2017-09-11 #382 @arfon
@raamana - your submission is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00382 ⚡️ 🚀 💥
